zanata_stat: Require to specify project and version
As described in bug 1670640, the current vesion of zanata_stat
only honors stats from a single combination of project and version.
To avoid this situation, this commit forces users to specify
a single branch (Zanata version) and a single project.

A further patch can improve the tool to merge statistics
from all versions and projects.

Change-Id: I806941e264747b8116ff4c0881e26cf890e516a8
Partial-Bug: #1670640