Move python files into importable path
It is useful when adding unit tests.

Defines python package namespace 'openstack_i18n'.
Also install executable scritps as 'openstack-zanata-stats'
and 'openstack-zanata-users'.

translation_team.yaml is moved to data/ directory.
The old path is kept as symlink as it is used by Stackalytics.

Note that check_yaml_file.py is only for testing
and it is not moved in this patch.

Change-Id: I60208456da7358356c944a9882577a34c981a1f2